requirements.txt
torch
torchvision
streamlit
flask
flask-cors
matplotlib
pillow
scikit-learn
